Interfaces and <ThirdwebProvider/> abstraction
