+++ src/embeds/erc1155.tsx
@@ -1,24 +1,17 @@
-import { ChakraProvider, Flex, useColorMode } from "@chakra-ui/react";
-import { css, Global } from "@emotion/react";
-import { ThirdwebProvider, useContract, useNFT } from "@thirdweb-dev/react";
-import { ThirdwebStorage } from "@thirdweb-dev/storage";
+import { Flex, useColorMode } from "@chakra-ui/react";
+import { useContract, useNFT } from "@thirdweb-dev/react";
 import React, { useEffect } from "react";
 import { createRoot } from "react-dom/client";
 import { Body } from "src/shared/body";
 import { ERC1155ClaimButton } from "src/shared/claim-button-erc1155";
 import { Footer } from "../shared/footer";
 import { Header } from "../shared/header";
-import { useGasless } from "../shared/hooks/useGasless";
-import chakraTheme from "../shared/theme";
-import { fontsizeCss } from "../shared/theme/typography";
 import { TokenMetadataPage } from "../shared/token-metadata-page";
-import { Chain, getChainBySlug } from "@thirdweb-dev/chains";
+import AppLayout from "src/shared/app-layout";
+import { BaseEmbedProps } from "src/shared/types/base";
 
-interface Erc1155EmbedProps {
-  contractAddress: string;
+interface Erc1155EmbedProps extends BaseEmbedProps {
   tokenId: string;
-  colorScheme: "light" | "dark";
-  primaryColor: string;
 }
 
 const Erc1155Embed: React.FC<Erc1155EmbedProps> = ({
@@ -69,61 +62,19 @@ const Erc1155Embed: React.FC<Erc1155EmbedProps> = ({
 const urlParams = new URL(window.location.toString()).searchParams;
 
 const App: React.FC = () => {
-  const chain =
-    urlParams.get("chain") && urlParams.get("chain")?.startsWith("{")
-      ? JSON.parse(String(urlParams.get("chain")))
-      : urlParams.get("chain");
-  const tempChain = getChainBySlug(
-    typeof chain === "string" ? chain : chain.slug,
-  );
-  const activeChain: Chain | string =
-    typeof chain === "string" ? chain : { ...chain, icon: tempChain.icon };
   const contractAddress = urlParams.get("contract") || "";
   const tokenId = urlParams.get("tokenId") || "0";
-  const relayerUrl = urlParams.get("relayUrl") || "";
-  const biconomyApiKey = urlParams.get("biconomyApiKey") || "";
-  const biconomyApiId = urlParams.get("biconomyApiId") || "";
-
   const colorScheme = urlParams.get("theme") === "dark" ? "dark" : "light";
   const primaryColor = urlParams.get("primaryColor") || "purple";
-
-  const ipfsGateway = urlParams.get("ipfsGateway");
-
-  const sdkOptions = useGasless(relayerUrl, biconomyApiKey, biconomyApiId);
-
   return (
-    <>
-      <Global
-        styles={css`
-          :host,
-          :root {
-            ${fontsizeCss};
-          }
-        `}
+    <AppLayout urlParams={urlParams}>
+      <Erc1155Embed
+        contractAddress={contractAddress}
+        tokenId={tokenId}
+        colorScheme={colorScheme}
+        primaryColor={primaryColor}
       />
-      <ChakraProvider theme={chakraTheme}>
-        <ThirdwebProvider
-          activeChain={activeChain}
-          sdkOptions={sdkOptions}
-          storageInterface={
-            ipfsGateway
-              ? new ThirdwebStorage({
-                  gatewayUrls: {
-                    "ipfs://": [ipfsGateway],
-                  },
-                })
-              : undefined
-          }
-        >
-          <Erc1155Embed
-            contractAddress={contractAddress}
-            tokenId={tokenId}
-            colorScheme={colorScheme}
-            primaryColor={primaryColor}
-          />
-        </ThirdwebProvider>
-      </ChakraProvider>
-    </>
+    </AppLayout>
   );
 };
