+++ src/shared/claim-button-erc20.tsx
@@ -27,11 +27,10 @@ import React, { useMemo, useState } from "react";
 import { parseIneligibility } from "../utils/parseIneligibility";
 import { useDebounce } from "./hooks/useDebounce";
 import chakraTheme from "./theme";
+import { ThemeProps } from "./types/base";
 
-interface ClaimButtonProps {
+interface ClaimButtonProps extends ThemeProps {
   contract?: Exclude<DropContract, EditionDrop | SignatureDrop | NFTDrop>;
-  primaryColor: string;
-  colorScheme: "light" | "dark";
 }
 
 export const ERC20ClaimButton: React.FC<ClaimButtonProps> = ({
