+++ src/shared/claim-button-erc721.tsx
@@ -28,11 +28,10 @@ import React, { useMemo, useState } from "react";
 import { parseIneligibility } from "../utils/parseIneligibility";
 import { useDebounce } from "./hooks/useDebounce";
 import chakraTheme from "./theme";
+import { ThemeProps } from "./types/base";
 
-interface ClaimButtonProps {
+interface ClaimButtonProps extends ThemeProps {
   contract?: Exclude<DropContract, TokenDrop | EditionDrop>;
-  primaryColor: string;
-  colorScheme: "light" | "dark";
 }
 
 export const ERC721ClaimButton: React.FC<ClaimButtonProps> = ({
