+++ src/shared/types/base.ts
@@ -1,0 +1,8 @@
+export interface ThemeProps {
+  colorScheme: "light" | "dark";
+  primaryColor: string;
+}
+
+export interface BaseEmbedProps extends ThemeProps {
+  contractAddress: string;
+}
